В кейсе 1 вынес функции чтобы их можно было импортировать, но не могу уже понять где ошибка. В итоге почему-то выходит только 15 строк =/
